louis/requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
Pillow==5.1.0
pkg-resources==0.0.0
python-decouple==3.1
pytz==2018.4
